requirements.txt
# Core dependencies
playwright==1.41.0
openai==1.12.0
python-dotenv==1.0.0

# Utilities
beautifulsoup4==4.12.3
html5lib==1.1

# Logging and debugging
colorama==0.4.6
rich==13.7.0
